// src/model/Auto.ts
export interface Auto {
    marca: string;
    modelo: string;
    anio: number;
    patente: string;
    color: string;
    numeroChasis: string;
    motor: string;
    duenio: string;
}

export type AutoListing = {
    marca: string;
    modelo: string;
    patente: string;
    duenio: string;
};